feat:test: unit-tests for all kinka versions (min, dev, orig); fix: production-js-loader
