+++ lib/kinka.js
@@ -10,7 +10,7 @@
 var helpers = require('./helpers')
 var isUndefined = helpers.isUndefined
 
-// <dev-version-code>
+// <dev-code>
 if (process.env.NODE_ENV !== 'production') {
   var globalObject = typeof global !== 'undefined' ? global : window
   /**
@@ -69,7 +69,7 @@ if (process.env.NODE_ENV !== 'production') {
     }
   }
 }
-// </dev-version-code>
+// </dev-code>
 
 /**
  * Create new instance of Kinka
@@ -99,15 +99,15 @@ function Kinka(instanceOptions) {
     typeCheck(
       customMethods,
       'array',
-      'custom methods with name in your kinka instance'
+      'option "customMethods" in your kinka instance'
     )
     for (var i in customMethods) {
       typeCheck(
         customMethods[i],
         'string',
-        'custom method with name ' +
+        'custom method with name "' +
           customMethods[i] +
-          ' in your kinka instance option customMethods'
+          '" in your kinka instance option customMethods'
       )
       this[customMethods[i]] = createMethodHandler(customMethods[i])
     }
@@ -148,7 +148,7 @@ function Kinka(instanceOptions) {
 
 function createMethodHandler(methodName, hasBody) {
   return function(path, optionsOrData, data) {
-    // <dev-version-code>
+    // <dev-code>
     if (!hasBody && typeof data !== 'undefined') {
       console.warn(
         '[kinka warning] Your data for the request ' +
@@ -158,7 +158,7 @@ function createMethodHandler(methodName, hasBody) {
           ' will not be pasted because third argument is specifically for PUT, POST, PATCH methods. Please use property "data" in second argument. \r\n Example kinka.head(\'/all\', { data: myData })'
       )
     }
-    // </dev-version-code>
+    // </dev-code>
     var config = hasBody ? [data, optionsOrData] : [optionsOrData]
     return this.custom(methodName, path, config[0], config[1])
   }
